Add support to subscribe to keyboard events (#165)
This commit is not strictly necessary, once lib users can make their own subscribe
to `Keyboard` events. On the other hand, this helps everything related to keyboard
to be on the same place.

Also, this commit makes the lib compliant with its documentation, which state that
`onKeyboardWillShow` is a possible prop for the component.